iTunes Style Getter improvements
Changed it to write info in the format "Genre/Style". Added comments. Also, split up the genre, style, and URL queries into three separate parts, using flags (arguments to the Ruby script) to switch between them.

To query for a genre, use the 'g' flag. For a style, use the 's' flag. For a URL, use any or no flag.