newsboat: add pocket-save script, fix macro
The new version of newsboat made it a startup error if a macro doesn't
contain an operation, and grouped commands (with {}) aren't interpreted
as operations. So if I want to background and disown the whole
save-to-pocket macro, I have to make it an external script and call that
from newsboat.


Former-commit-id: beb650361c856d84a249ad8fa8fea9a921b059e0