setbg & set-theme: play nicer with Alacritty
I added a variant of alacritty-color-export into setbg, so that when
setting a new background, a color scheme based on that background and
the current theme (light/dark) will also get written into alacritty.yml.
That means that in set-theme, I only need to change the 'colors: *dark'
to 'colors: *light' and vice-versa.