Attempt to unify the way links are handled
I basically want to have a menu pop up with possible actions. I want it
to work with fzf on mac, and dmenu on linux, depending on the CHOOSER
variable. Also it should work with a plumbing keybinding. This is the
first step, the system will be improved in later commits.